+++ src/nd.hpp
@@ -180,6 +180,8 @@ public:
 	// Reductions.
 	static Ref<NDArray> sum(const Variant& a, const Variant& axes);
 	static Ref<NDArray> prod(const Variant& a, const Variant& axes);
+	static Ref<NDArray> cumsum(const Variant& a, const Variant& axis);
+	static Ref<NDArray> cumprod(const Variant& a, const Variant& axis);
 	static Ref<NDArray> mean(const Variant& a, const Variant& axes);
 	static Ref<NDArray> median(const Variant& a, const Variant& axes);
 	static Ref<NDArray> variance(const Variant& a, const Variant& axes);